Move sample to plugins (#130)
* Move sample under plugins

* Add entry_point for sample

* Clean and fix tests

* Clean up and add tests

* Update docs pages
